+++ source/MathFighterXNA/MathFighterXNA/Entity/Menu/MenuItem.cs
@@ -66,20 +66,20 @@ namespace ClownSchool.Entity.Menu {
 
         public override void Draw(Microsoft.Xna.Framework.Graphics.SpriteBatch spriteBatch) {
             if (RenderBalloons) {
-                spriteBatch.Draw(Assets.MenuBalloons, new Rectangle(X - 55, Y - 190, 150, 195), Color.White);
-                spriteBatch.Draw(Assets.MenuBalloons, new Rectangle(X + 123, Y - 190, 150, 195), Color.White);
+                spriteBatch.Draw(Assets.MenuBalloons, new Rectangle((int)X - 55, (int)Y - 190, 150, 195), Color.White);
+                spriteBatch.Draw(Assets.MenuBalloons, new Rectangle((int)X + 123, (int)Y - 190, 150, 195), Color.White);
             }
 
             if (RenderRopes) {
-                spriteBatch.Draw(Assets.MenuRope, new Rectangle(X + 10, Y + 95, 3, 115), Color.White);
-                spriteBatch.Draw(Assets.MenuRope, new Rectangle(X + 188, Y + 95, 3, 115), Color.White);
+                spriteBatch.Draw(Assets.MenuRope, new Rectangle((int)X + 10, (int)Y + 95, 3, 115), Color.White);
+                spriteBatch.Draw(Assets.MenuRope, new Rectangle((int)X + 188, (int)Y + 95, 3, 115), Color.White);
             }
 
             if (selected) {
-                spriteBatch.Draw(Assets.MenuSignGlow, new Rectangle(X - 30, Y - 30, Size.X + 60, Size.Y + 60), Color.White);
+                spriteBatch.Draw(Assets.MenuSignGlow, new Rectangle((int)X - 30, (int)Y - 30, Size.X + 60, Size.Y + 60), Color.White);
             }
-            
-            spriteBatch.Draw(Graphic, new Rectangle(X, Y, Size.X, Size.Y), Color.White);
+
+            spriteBatch.Draw(Graphic, new Rectangle((int)X, (int)Y, Size.X, Size.Y), Color.White);
 
             if (hoverTime > 0 && hoverTime <= maxHoverTime) {
                 PlayerHand hand = (PlayerHand)GetFirstCollidingEntity("hand");
@@ -87,7 +87,7 @@ namespace ClownSchool.Entity.Menu {
                     return;
 
                 for (int i = 0; i <= 360; i++) {
-                    var destRect = new Rectangle(hand.X - 50, hand.Y - 50, 1, 20);
+                    var destRect = new Rectangle((int)hand.X - 50, (int)hand.Y - 50, 1, 20);
 
                     var asset = Assets.CirclePartFilled;
                     if ((360 / maxHoverTime) * hoverTime <= i) {
