+++ source/MathFighterXNA/MathFighterXNA/Player/TutorialHand.cs
@@ -26,7 +26,7 @@ namespace ClownSchool {
         public override void Draw(Microsoft.Xna.Framework.Graphics.SpriteBatch spriteBatch) {
             var glove = IsGrabbing ? Assets.GloveFist : Assets.Glove;
 
-            spriteBatch.Draw(IsGrabbing ? Assets.GloveFist : Assets.Glove, new Rectangle(X, Y, 56, 64), null, Color.White, 0, new Vector2(glove.Width / 2, glove.Height / 2), Hand == JointType.HandLeft ? SpriteEffects.None : SpriteEffects.FlipHorizontally, 0);      
+            spriteBatch.Draw(IsGrabbing ? Assets.GloveFist : Assets.Glove, new Rectangle((int)X, (int)Y, 56, 64), null, Color.White, 0, new Vector2(glove.Width / 2, glove.Height / 2), Hand == JointType.HandLeft ? SpriteEffects.None : SpriteEffects.FlipHorizontally, 0);      
         }
     }
 }
